requirements.txt
uvicorn==0.35.0
fastmcp==2.12.3
pyyaml==6.0.2
httpx==0.28.1
pydantic==2.11.7
